requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
python-dotenv==1.0.1
pydantic[email]==2.9.0
pydantic-settings==2.5.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
PyJWT==2.8.1
email-validator==2.1.0
